backend/requirements.txt
fastapi
uvicorn
paho-mqtt
pymongo
python-dotenv
joblib
scikit-learn 